% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_day.R
\name{get_last_x_games_per_team}
\alias{get_last_x_games_per_team}
\title{Récupération des x derniers matchs par équipe}
\usage{
get_last_x_games_per_team(game_logs, nb_last_games)
}
\arguments{
\item{game_logs}{(dfr) la table complète des stats par joueur/match}

\item{nb_last_games}{(int) le nombre de matchs souhaités}
}
\value{
(dfr) table des nameTeam et dateGame des x derniers matchs
}
\description{
Récupération des x derniers matchs par équipe
}
