% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_day.R
\name{get_match_day}
\alias{get_match_day}
\title{Récupération des matchs d'un jour souhaité}
\usage{
get_match_day(game_dates = Sys.Date())
}
\arguments{
\item{game_dates}{les dates des matchs souhaités}
}
\value{
(dfr) table des matchs des dates souhaitées
}
\description{
Récupération des matchs d'un jour souhaité
}
\examples{
# get_match_day()
}
