% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_day.R
\name{get_nb_pts_seuil}
\alias{get_nb_pts_seuil}
\title{Calculer le nombre de fois où le joueur a scoré + qu'un seuil}
\usage{
get_nb_pts_seuil(pts_last_games_long, seuil_pts)
}
\arguments{
\item{pts_last_games_long}{(dfr) la table des points sur les x derniers matchs au format long}

\item{seuil_pts}{le nombre de points seuil}
}
\description{
Calculer le nombre de fois où le joueur a scoré + qu'un seuil
}
