% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_day.R
\name{get_pts_last_games_long}
\alias{get_pts_last_games_long}
\title{Récupérer les points des x derniers matchs, format long}
\usage{
get_pts_last_games_long(game_data, nb_last_games)
}
\arguments{
\item{game_data}{(dfr) la table complète des stats par joueur/match}

\item{nb_last_games}{(int) le nombre de matchs souhaités}
}
\description{
Récupérer les points des x derniers matchs, format long
}
