% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_year.R
\name{name_year}
\alias{name_year}
\title{A Function to Find the mean of 2P of NBA Player by Year}
\usage{
name_year(x)
}
\description{
This function allows you to Find the mean of 2P of NBA Player in a specific Year
}
\examples{
name_year()
}
\keyword{2P}
\keyword{nba,mean}
\keyword{of}
