% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_year.R
\name{oldest_year}
\alias{oldest_year}
\title{A Function to Find Oldest NBA Player by Year}
\usage{
oldest_year(x)
}
\description{
This function allows you to Find the oldest NBA Player in a specific Year
}
\examples{
oldest_year()

}
\keyword{nba,}
\keyword{oldest}
